+++ tavernaSender.php
@@ -1,15 +1,28 @@
 <?php
       require_once 'sender.php';
       require_once 'TavernaException.php';
+      require_once 'TavernaCurlConnection.php';
                                   
 class TavernaSender extends Sender
-{
-/*
-* intitiallize the
-*/
+{ 
+
+  public $curl_connect;
+  
   function __construct($protocol,$hostname=null,$port=null,$context, $username=null,$password =null)
   {
     parent::__construct($protocol.'://'.$hostname.":".$port .'/'. $context."/rest/runs/",$password,$username);
+    $this->curl_connect =new TavernaCurlConnection();
+    $this->curl_connect->username = $this->username;
+    $this->curl_connect->password = $this->password;
+  }
+
+
+  function set_ssl()
+  {
+	
+    $this->curl_connect->sslVersion = 3;
+    $this->curl_connect->verifyHost = FALSE;
+    $this->curl_connect->verifyPeer = FALSE; 
   }
   
   //would like to update this function to use tuque HttpConnection class
@@ -17,31 +30,16 @@ class TavernaSender extends Sender
   {
       if (!empty($message))
       {
-          $ch = curl_init();
-          curl_setopt($ch, CURLOPT_URL, $this->hostname);
-          curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-          curl_setopt($ch, CURLOPT_POST, true);
-          curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, FALSE);
-          curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, FALSE);
-          curl_setopt($ch, CURLOPT_SSLVERSION, 3);
-          curl_setopt($ch, CURLOPT_HEADER, true);
-          curl_setopt($ch, CURLOPT_USERPWD, "$this->username:$this->password");
-          curl_setopt($ch, CURLOPT_HTTPHEADER, array('Content-Type: application/vnd.taverna.t2flow+xml'));
-          curl_setopt($ch, CURLOPT_POSTFIELDS, $message);
-          curl_setopt($ch, CURLOPT_HTTPAUTH, CURLAUTH_BASIC);
-          $content = curl_exec($ch);
-          $code = curl_getinfo($ch,CURLINFO_HTTP_CODE);
-          print('taverna host = '.$this->hostname);
-          print("post code = ".$code);
-          print('username = '. $this->username.' password '.$this->password);
-          if ($code != 201)
-          {
-              curl_close($ch);
-              throw new TavernaException($content);
-          }
-          
-          curl_close($ch);
-          return $content;
+
+	 $this->set_ssl();
+	 $response = $this->curl_connect->postRequest($this->hostname,'string',$message,'application/vnd.taverna.t2flow+xml');
+        
+	 echo $response['content'];
+	 if($response['status'] !=201)
+         { 
+		throw new TavernaException($response['headers'].$response['content']);
+	 }
+	 return $response['headers'].$response['content'];
       }
       
       return null;
@@ -58,62 +56,50 @@ class TavernaSender extends Sender
     
     return null;
   }
-  
+
+  //HttpConnection:putRequest  
   function run_t2flow($uuid)
   {
       if (!empty($uuid))
       {
-          $ch = curl_init();
-          curl_setopt($ch, CURLOPT_URL, $this->hostname.$uuid."/status/");
-          curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);    
-          curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, FALSE);
-          curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, FALSE);
-          curl_setopt($ch, CURLOPT_SSLVERSION, 3);
-          curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "PUT");
-          curl_setopt($ch, CURLOPT_HEADER, true);
-          curl_setopt($ch, CURLOPT_USERPWD, "$this->username:$this->password");
-          curl_setopt($ch, CURLOPT_HTTPHEADER, array('Content-Type: text/plain'));
-          curl_setopt($ch, CURLOPT_POSTFIELDS, "Operating");
-          curl_setopt($ch, CURLOPT_HTTPAUTH, CURLAUTH_BASIC);
-          $result = curl_exec($ch);
-      
-          if (curl_getinfo($ch,CURLINFO_HTTP_CODE) != 200)
-          {
-              curl_close($ch);
-              throw new TavernaException($result);
-          }
-      
-      curl_close($ch);
-      return $result;
+
+        $url= $this->hostname.$uuid.'/status/';
+
+	$this->set_ssl(); 
+
+	$response = $this->curl_connect->tavernaPutRequest($url,'string','Operating','text/plain');
+
+	if($response['status'] != 200)
+        { 
+		throw new TavernaException($response['headers'].$response['content']);
+	}
+
+	
+	return $response['headers'].$response['content'];
+ 
       }
       return null;
-  }
+ }
   
+
+  //HttpConnection:getRequest
   function get_status($uuid)
   {
       if (!empty($uuid))
       {
-          $ch = curl_init();
-          curl_setopt($ch, CURLOPT_URL, $this->hostname.$uuid."/status/");
-          curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);                   
-          curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "GET");
-          curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, FALSE);
-          curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, FALSE);
-          curl_setopt($ch, CURLOPT_SSLVERSION, 3);
-          curl_setopt($ch, CURLOPT_HEADER, true);
-          curl_setopt($ch, CURLOPT_USERPWD, "$this->username:$this->password");
-          curl_setopt($ch, CURLOPT_HTTPHEADER, array('Content-Type: text/plain'));
-          curl_setopt($ch, CURLOPT_HTTPAUTH, CURLAUTH_BASIC);
-          $result = curl_exec($ch);
-      
-          if (curl_getinfo($ch,CURLINFO_HTTP_CODE) != 200)
-          {
-              curl_close($ch);
-              throw new TavernaException($result);
-          }
-      
-          curl_close($ch);
+
+	 $url=$this->hostname.$uuid."/status/";
+	 $this->set_ssl();
+	 $response = $this->curl_connect->getRequest($url,FALSE,NULL);
+        
+	 if($response['status'] !=200)
+         { 
+		throw new TavernaException($response['headers'].$response['content']);
+	 }
+	
+	 return $response['headers'].$response['content'];
       }
+
       return null;
   }
   
@@ -121,60 +107,42 @@ class TavernaSender extends Sender
   {
     if (!empty($uuid))
     {
-        $ch = curl_init();
-        curl_setopt($ch, CURLOPT_URL, $this->hostname.'/'.$uuid.'/');
-        curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);                   
-        curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "DELETE");
-        curl_setopt($ch, CURLOPT_HEADER, true);
-        curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, FALSE);
-        curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, FALSE);
-        curl_setopt($ch, CURLOPT_SSLVERSION, 3);
-        curl_setopt($ch, CURLOPT_USERPWD, $this->username.':'.$this->password);
-        curl_setopt($ch, CURLOPT_HTTPHEADER, array('text/plain'));
-        $result = curl_exec($ch);
-    
-        if (curl_getinfo($ch,CURLINFO_HTTP_CODE) != 202)
-        {
-            curl_close($ch);
-            throw new TavernaException($result);
-        }
-    
-        curl_close($ch);
-        return $result;
+
+	$url= $this->hostname.$uuid.'/';
+
+        $this->set_ssl();
+	$response = $this->curl_connect->deleteRequest($url);
+
+	if($response['status'] !=202)
+        { 
+		throw new TavernaException($response['headers'].$response['content']);
+	}
+ 	return $response['headers'].$response['content'];
     }
     
     return null;
-  }
+  } 
+
   //would like to use tuque here
   function add_input($uuid, $key, $value)
   {
       if (!empty($uuid) && !empty($key) && !empty($value))
       {
+
           $input = '<t2sr:runInput xmlns:t2sr="http://ns.taverna.org.uk/2010/xml/server/rest/">
                       <t2sr:value>'.$value.'</t2sr:value>             
                     </t2sr:runInput>';
-          $ch = curl_init();
-          curl_setopt($ch, CURLOPT_URL, $this->hostname.$uuid."/input/input/".$key);
-          curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);                   
-          curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "PUT");
-          curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, FALSE);
-          curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, FALSE);
-          curl_setopt($ch, CURLOPT_SSLVERSION, 3);
-          curl_setopt($ch, CURLOPT_HEADER, true);
-          curl_setopt($ch, CURLOPT_USERPWD, "$this->username:$this->password");
-          curl_setopt($ch, CURLOPT_HTTPHEADER, array('Content-Type: application/xml'));
-          curl_setopt($ch, CURLOPT_POSTFIELDS, $input);
-          curl_setopt($ch, CURLOPT_HTTPAUTH, CURLAUTH_BASIC);
-          $result = curl_exec($ch);
-          
-          if (curl_getinfo($ch,CURLINFO_HTTP_CODE) != 200)
-          {
-              curl_close($ch);
-              throw new TavernaException($result);
-          }
-          
-          curl_close($ch);
-          return $result;
+	 
+	 $url=$this->hostname.$uuid."/input/input/".$key;
+
+	 $response = $this->curl_connect->tavernaPutRequest($url,'string',$input,'application/xml');
+        
+	 if($response['status'] !=200)
+         { 
+		throw new TavernaException($response['headers'].$response['content']);
+	 }
+
+	 return $response['headers'].$response['content'];
       }
       
       return null;
